Fix flaky acked-input bridge test by removing buffer-dependent blocking send
The ackedInputModeDrainsOnAck test blocked on a 5 MiB sendBlocking (5s
semaphore timeout on NWConnection's .contentProcessed) before emitting
any acks, while RemotePTYBridgeSession intentionally pauses socket
reads at its 4 MiB acked-input window. Whether the ~1 MiB tail was
accepted in time depended on kernel socket buffer autotuning, not on
the code under test: a probe with no acks ever emitted saw the 5 MiB
completion fire in 6 ms (tail absorbed by buffers) while a 16 MiB
completion never fired in 8 s with the server parked exactly at the
4,194,304-byte cap. The assertion could pass with a broken server and
fail with a correct one, which is how it turned main red in
https://github.com/manaflow-ai/cmux/actions/runs/29075799069.

Test-only fix: send the 5 MiB asynchronously via a new tracked send
that returns a pollable completion flag, and assert the flag only
after the ack-drain phase, where completion is deterministic. All
pre-existing window/seq/delivery assertions are preserved verbatim;
sendBlocking is deleted. BridgeTestClient moves to its own file so the
test file (was 499 lines) stays under the 500-line budget threshold.

Verified with 40/40 green focused iterations and 2/2 full-package
swift test runs after the change.
